Time: 0 ms (100.00%) | Memory: 7.9 MB (1.48%) - LeetSync

+++ 371-sum-of-two-integers/sum-of-two-integers.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     int getSum(int a, int b) {
-        while (b != 0) {
-            unsigned int carry = (unsigned int)(a & b) << 1;
-            a = a ^ b;
-            b = carry;
-        }
-        return a;
+          while (b != 0) {
+        int carry = (a & b) << 1;
+        a = a ^ b;
+        b = carry;
+    }
+    return a;
     }
 };
